Support long milliseconds
`1n` to ms is 2592000000ms, which is greater than max int at 2**31-1. This allows any acceptable length of ms to be parsed.

+++ src/utils/DateTime.java
@@ -169,7 +169,7 @@ public class DateTime {
       unit++;
     }
     try {
-      interval = Integer.parseInt(duration.substring(0, unit));
+      interval = Long.parseLong(duration.substring(0, unit));
     } catch (NumberFormatException e) {
       throw new IllegalArgumentException("Invalid duration (number): " + duration);
     }
